Major Change in Kiosk, Register and Store class.
Major Change in Kiosk, Register and Store class. Orders are created via
Kiosks and Registers, and Orders can be created by a customer if he is
sitting on a kiosk according to business rule. Removing createOrder()
method in store and adding createOrderViaKiosk() method and
createOrderViaRegister() method in Store, and making subsequent changes
in the related classes.