Implemented Free pizza certificate system
Successfully implemented and tested free pizza certificate system based
on rewards point. Default rewards point set to 150 for free pizza
certificate. This can be changes by the manager. The manager UI for
changing the certain defined amount for getting a free pizza (the
minimum points required is not set yet).

By using setRequiredPointForFreePizzaCertificate(int point), this UI can
be built. We have to build a UI with only manager access which calls
this function with the value the manager sets. The internal records will
be updated accordingly.

Frequent buyer account is implemented internally. A UI for it has to be
developed.