Created a order activity to take in orders.
Created a order activity to take in the orders. This order activity is
going to read the store json object from the server from the /store
address in the server, and going to populate the list of available items
and menus. Peter is going to take care of this part from now on. The
rest of the things in the android app is implemented.